Removing duplicate type definition.
