+++ tile_engine/ops/gemm/benchmark_gemm.hpp
@@ -13,11 +13,8 @@
 #include "ck_tile/host.hpp"
 #include "gemm_common.hpp"
 
-// Define the actual types used by the kernels
-using ADataType   = ck_tile::half_t;
-using BDataType   = ck_tile::half_t;
-using AccDataType = float;
-using CDataType   = ck_tile::half_t;
+// Data types are defined by the generated kernel headers
+// No hardcoded type definitions here to avoid conflicts
 
 using ALayout = ck_tile::tensor_layout::gemm::RowMajor;
 using BLayout = ck_tile::tensor_layout::gemm::ColumnMajor;
@@ -151,65 +148,6 @@ inline std::string get_rocm_version()
         return version;
     }
     return "Unknown";
-}
-
-// Helper function to determine if a layout is row-major
-template <typename Layout>
-constexpr auto is_row_major(Layout)
-{
-    return ck_tile::bool_constant<std::is_same_v<Layout, ck_tile::tensor_layout::gemm::RowMajor>>{};
-}
-
-// Permutation function for pk_int4_t
-template <typename Tensor>
-void permute_vectors_i4x4_b(Tensor& tensor)
-{
-    const ck_tile::index_t K = tensor.get_length(0);
-    const ck_tile::index_t N = tensor.get_length(1);
-    // vector pk_i4x4 permute
-    for(int i = 0; i < N; i++)
-    {
-        for(int j = 0; j < K; j += 8)
-        {
-            int8_t input[8];
-
-            for(int k = 0; k < 4; k++)
-            {
-                int8_t i4x2      = tensor(j + k * 2, i).data;
-                input[k * 2 + 0] = (i4x2 >> 4) & 0xf;
-                input[k * 2 + 1] = (i4x2 >> 0) & 0xf;
-            }
-
-            // permute 01234567->20643175
-            {
-                int8_t hi        = input[2];
-                int8_t lo        = input[0];
-                int8_t i4x2      = (hi << 4) | lo;
-                tensor(j + 0, i) = i4x2;
-            }
-
-            {
-                int8_t hi        = input[6];
-                int8_t lo        = input[4];
-                int8_t i4x2      = (hi << 4) | lo;
-                tensor(j + 2, i) = i4x2;
-            }
-
-            {
-                int8_t hi        = input[3];
-                int8_t lo        = input[1];
-                int8_t i4x2      = (hi << 4) | lo;
-                tensor(j + 4, i) = i4x2;
-            }
-
-            {
-                int8_t hi        = input[7];
-                int8_t lo        = input[5];
-                int8_t i4x2      = (hi << 4) | lo;
-                tensor(j + 6, i) = i4x2;
-            }
-        }
-    }
 }
 
 template <typename ADataType, typename BDataType, typename AccDataType, typename CDataType>
